レイアウト調整途中

+++ resources/views/posts/index.blade.php
@@ -24,12 +24,23 @@
 <!-- ツイート表示 -->
 <div class="post-wrapper">
                 @foreach($posts as $post)
-                <div>
-                  <p>{{$post->user_id}}</p>
-                  <p>{{$post->post}}</p>
-                  <p>{{$post->created_at}}</p>
-                  <label class="js-modal-open" post="{{ $post->post }}" post_id="{{ $post->id }}">編集</label>
-                  <a href="/delete/{{$post->id}}" onclick="return confirm('こちらの投稿を削除してもよろしいでしょうか？')">削除</a>
+                <div class=posts>
+                    <div class=post-box>
+                    <img src="{{ asset('storage/profiles/'.$post->image) }}">
+                    <div post-content>
+                      <div class=n-c-box>
+                        <p class=post-name>{{$post->user_id}}</p>
+                        <p class=post-created_at>{{$post->created_at}}</p>
+                      </div>
+                        <p class=post-post>{{$post->post}}</p>
+                      </div>
+                    </div>
+                    
+                  <div class=u-d-box>
+                  <input type = "image" src = "../images/edit.png" post="{{ $post->post }}" post_id="{{ $post->id }}" alt = "編集" class =js-modal-open>
+                  <input type = "image" src = "../images/trash-h.png" action="/delete/{{$post->id}}" onclick="return confirm('こちらの投稿を削除してもよろしいでしょうか？')" alt = "削除" class =post-trash>
+                    <a href="/delete/{{$post->id}}" onclick="return confirm('こちらの投稿を削除してもよろしいでしょうか？')">削除</a>
+                  </div>
                 </div>
                 @endforeach
